Fix compile warnings
git-svn-id: https://svn.cern.ch/reps/atlasperf/CombPerf/FlavorTag/JetTagPerformanceCalibration/Combination/trunk@252301 47dd0619-b7a8-462c-89d8-56395fd97e59

+++ src/CDIConverter.cxx
@@ -176,7 +176,7 @@ namespace {
   {
     for (unsigned int ibin = 0; ibin < ana.bins.size(); ibin++) {
       const CalibrationBin &bin (ana.bins[ibin]);
-      for (int i_sys = 0; i_sys < bin.systematicErrors.size(); i_sys++) {
+      for (size_t i_sys = 0; i_sys < bin.systematicErrors.size(); i_sys++) {
 	const SystematicError &e(bin.systematicErrors[i_sys]);
 	if (e.name == ename) {
 	  return e.uncorrelated;
